RPC, cosmetic: Create rpcmining.cpp as new home for mining-related RPC code
