+++ src/makefile.mingw
@@ -58,6 +58,7 @@ OBJS= \
     obj/bitcoinrpc.o \
     obj/rpcdump.o \
     obj/rpcnet.o \
+    obj/rpcmining.o \
     obj/rpcrawtransaction.o \
     obj/script.o \
     obj/sync.o \
