+++ src/makefile.osx
@@ -85,6 +85,7 @@ OBJS= \
     obj/bitcoinrpc.o \
     obj/rpcdump.o \
     obj/rpcnet.o \
+    obj/rpcmining.o \
     obj/rpcrawtransaction.o \
     obj/script.o \
     obj/sync.o \
